Refactor: Move `ItemTitle` to its own file
