+++ crates/viewer/re_selection_panel/src/lib.rs
@@ -1,6 +1,7 @@
 //! The UI for the selection panel.
 
 mod defaults_ui;
+mod item_title;
 mod selection_history_ui;
 mod selection_panel;
 mod space_view_entity_picker;
@@ -8,6 +9,7 @@ mod space_view_space_origin_ui;
 mod visible_time_range_ui;
 mod visualizer_ui;
 
+pub use item_title::ItemTitle;
 pub use selection_panel::SelectionPanel;
 
 #[cfg(test)]
